Cleanup code bash script
Cleanup Code
This is a command-line tool for cleaning up code using JetBrains' cleanupcode tool.

Prerequisites
JetBrains' cleanupcode tool must be installed and available in the system's PATH.

Usage
Run the script with:

Copy code
sh ./scripts/git/cleanupcode.sh

The script will do the following:
Check if there are any unstaged changes in the repository. If there are, it will ask the user to commit them before running the script.
Check if there are any staged, uncommitted changes in the repository. If there are, it will ask the user to commit them before running the script.
Run the cleanupcode tool on the specified solution file (Blef.sln in this case) with the specified profile (Blef: Full Cleanup) and disable the specified settings layers (SolutionPersonal).
If any files were reformatted, stage them for commit and create a commit with the message "Re-format code by JetBrains CleanupCode Tool".

Notes
The script is currently hardcoded to use Blef.sln as the solution file and Blef: Full Cleanup as the profile. These can be modified by editing the script.
The cleanupcode tool's verbosity is set to WARN in the script. This can be changed by modifying the --verbosity flag.